refactor(src): group modules into domain folders with import rewrite
