+++ src/architecture/doi-index.ts
