+++ src/architecture/prompt-trigger.ts
