+++ src/display/tex-display.ts
