+++ src/scale/library-scale.ts
