+++ src/ui/citation-card-layout.ts
