+++ src/ui/citation-popup.ts
