+++ src/ui/pin-registry.ts
